change pricePerNight to BigDecimal

+++ hotel-service/src/main/java/com/example/hotel_service/dto/RoomCreateRequest.java
@@ -1,6 +1,8 @@
 package com.example.hotel_service.dto;
 
+import java.math.BigDecimal;
+
 public record RoomCreateRequest(String description,
                                 Integer capacity,
-                                Integer pricePerNight
+                                BigDecimal pricePerNight
 ) {}
