+++ hotel-service/src/main/java/com/example/hotel_service/dto/RoomResponse.java
@@ -1,9 +1,11 @@
 package com.example.hotel_service.dto;
 
+import java.math.BigDecimal;
+
 public record RoomResponse(
         Long id,
         String description,
         Integer capacity,
-        Integer pricePerNight,
+        BigDecimal pricePerNight,
         Long hotelId
 ) {}
